% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot2_function.R
\name{boot2}
\alias{boot2}
\title{Bootstrapping with a progress bar}
\usage{
boot2(data, statistic, R, sim = "ordinary", stype = c("i", "f", "w"),
  strata = rep(1, n), L = NULL, m = 0, weights = NULL,
  ran.gen = function(d, p) d, mle = NULL, simple = FALSE, ...,
  parallel = c("no", "multicore", "snow"), ncpus = getOption("boot.ncpus",
  1L), cl = NULL)
}
\description{
This function reproduces the code from the \code{boot} package's
 \code{boot()} function, changing one line in order to add a progress
 bar using \code{pbapply} package. This solution was posted to
 StackOverflow by user psolymos and is shamelessly reproduced here.
 See http://goo.gl/F526BQ for details and \code{?boot::boot} for help.
}
