% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_estimate_parms_function.R
\name{boot_estimate_parms}
\alias{boot_estimate_parms}
\title{Bootstrapping with a progress bar}
\usage{
boot_estimate_parms(reps, scores, id.rate, nom.rate, multicore = FALSE)
}
\arguments{
\item{reps}{Number of bootstrapped repetitions.}

\item{scores}{Numeric vector of observed scores.}

\item{id.rate}{Identification rate. Must be between 0 and 1.}

\item{nom.rate}{Nomination rate. Must be between 0 and 1.}

\item{multicore.}{Logical value indicating whether parallel processing should be used.
Must be \code{TRUE} or \code{FALSE}. Defaults to \code{FALSE}.}
}
\description{
Bootstrapping with a progress bar
}
\examples{
# generate some data
set.seed(123)
scores <- r_identified(n=500, test.cutoff=.9, relyt=.93, valid=.5,
  nom.cutoff=.9)

# calculate the identification rate implied by the system parameters
id.rate <- marginal_psychometrics(test.cutoff=.9, relyt=.93, valid=.5,
  nom.cutoff=.9)$identification.rate

# calculate the nomination rate implied by the system parameters
nom.rate <- marginal_psychometrics(test.cutoff=.9, relyt=.93, valid=.5,
  nom.cutoff=.9)$nom.rate

# estimate the system parameters with 10 bootstrapped samples
boot_estimate_parms(scores, id.rate=id.rate, nom.rate=nom.rate, reps=10)

}
