% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_estimate_valid_function.R
\name{boot_estimate_valid}
\alias{boot_estimate_valid}
\title{Bootstrapping with a progress bar}
\usage{
boot_estimate_valid(reps, x, nom.rate, id.rate, pop.mean = 0,
  pop.sd = 1, adjust = 1)
}
\arguments{
\item{reps}{Number of bootstrapped repetitions.}

\item{x}{Numeric vector of observed scores.}

\item{nom.rate}{Nomination rate. Must be between 0 and 1.}

\item{id.rate}{Identification rate. Must be between 0 and 1.}

\item{pop.mean}{The known general population mean of x. Defaults to 0.}

\item{pop.sd}{The known general population standard deviation of x.
Defaults to 1.}

\item{adjust}{Controls the bandwidth of the density estimator. Defaults to 1.0, which
has been found to perform well in simulation.}
}
\description{
Bootstrapping with a progress bar
}
\examples{
# generate some observed scores
# (note the lack of a relyt argument)
# true validity is .6
set.seed(1)
x <- r_identified(n=500, test.cutoff=.9, valid=.6,
  nom.cutoff=.85)

# calculate the identification rate implied by the system parameters
id.rate <- marginal_psychometrics(test.cutoff=.9, valid=.6,
  nom.cutoff=.85)$identification.rate

# calculate the nomination rate implied by the system parameters
nom.rate <- marginal_psychometrics(test.cutoff=.9, valid=.6,
  nom.cutoff=.85)$nom.rate

# estimate the nomination validity with 10 bootstrapped samples
boot_estimate_valid(x, id.rate=id.rate, nom.rate=nom.rate, reps=10)

}
