% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_moments_function.R
\name{conditional_moments}
\alias{conditional_moments}
\title{Conditional moments of the observed score distribution given the true score
 and identification system parameters}
\usage{
conditional_moments(true.score, relyt, valid)
}
\arguments{
\item{true.score}{The student's true score on a standardized (z-score) metric.}

\item{relyt}{Confirmatory test reliability coefficient. Range (0, 1).
Must not be exactly 0 or 1.}

\item{valid}{Nomination validity coefficient. Controls the relatedness of the nomination
scores and the confirmatory test scores. Range (0, 1). Must not be exactly 0 or 1, and
must be less than the square root of the test reliability. If provided, the two-stage
version of the computation is performed.}
}
\description{
This function calculates the mean and variance of observed scores
 conditional on the true score and the identification system parameters. In a
 one-stage identification system, the only observed score is the observed test
 score, and so the mean and variance are scalar values. The distribution of
 observed scores is assumed to be normal.
}
\details{
In a two-stage identification system, there are observed scores for
 both the nomination and the confirmatory test. In this case, there is a mean vector
 and a covariance matrix, and the joint distribution is assumed to be bivariate
 normal. A two-stage system is one in which an initial nomination process is used
 to select students who are tested on the confirmatory assessment.

The observed score distribution for a student with a specific true score is justified
 by classical test theory, in which measurement errors are random variables. Because
 observed scores are true score (T) plus random error (E), the observed scores are
 also random variables.

If the argument valid is supplied, the two-stage conditional moments are
 returned. Otherwise the one-stage moments are returned.

This function returns a list containing the following:

conditional.mean: For one-stage systems, the mean of the confirmatory test scores.
 For two-stage systems, the mean vector of the bivariate normal distribution of
 the observed nomination and confirmatory test scores, in that order.

conditional.cov: For one-stage systems, the variance of the observed confirmatory
  test scores. For two-stage systems, the covariance matrix of the bivariate
  normal distribution of the observed nomination and confirmatory test scores,
  in that order.
}
\examples{

# one-stage system
conditional_moments(relyt=.9, true.score=2)

# two-stage system
conditional_moments(relyt=.9, true.score=2, valid=.6)
}
