% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_moments_1stage_function.R
\name{conditional_moments_1stage}
\alias{conditional_moments_1stage}
\title{Conditional mean and variance of the observed scores given
 the true score for a single-stage identification system}
\usage{
conditional_moments_1stage(relyt, true.score)
}
\arguments{
\item{relyt}{Confirmatory test reliability coefficient. Range (0, 1).
Must not be exactly 0 or 1.}

\item{true.score}{The student's true score on a standardized (z-score) metric.}
}
\description{
\code{conditional_moments_1stage} calculates the conditional mean and variance
 of the observed score (upon which identification decisions are based) given a
 student's true score, and the test reliability. Under classical test theory,
 the observed score X = T + E. Since E follows a normal distribution, the observed
 scores for a student with a particular true score are normally distributed
 around the true score, but regressed to the mean depending on the test reliability. This
 function returns the moments of f(X|T=t), where f() is the normal probability density
 function. One can imagine this as the idealized distribution of observed scores that
  would occur if a student with a given true score took the test an infinite number of times.
}
\details{
This function returns a list containing the following:

conditional.mean: The mean of the normal distribution of the observed scores.

conditional.var: The variance of the normal distribution of the observed scores.
}
\examples{
# test reliability is .9, and the student's
# true score is 2 SDs above the mean
#
conditional_moments_1stage(relyt=.9, true.score=2)

}
