% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_p_id_function.R
\name{conditional_p_id}
\alias{conditional_p_id}
\title{Conditional probability of identification given the true score}
\usage{
\code{conditional_p_id(true.score, relyt, test.cutoff, valid, nom.cutoff)}
}
\arguments{
\item{true.score}{The student's true score on a standardized (z-score) metric.}

\item{relyt}{Confirmatory test reliability coefficient. Range (0, 1).
Must not be exactly 0 or 1.}

\item{test.cutoff}{Confirmatory test cutoff percentile. Range (0, 1).
Must not be exactly 0 or 1. Can be a vector or scalar.}

\item{valid}{Nomination validity coefficient. Controls the relatedness of the nomination
scores and the confirmatory test scores. Range (0, 1). Must not be exactly 0 or 1, and
must be less than the square root of the test reliability.}

\item{nom.cutoff}{Nomination cutoff percentile. Range (0, 1).
Must not be exactly 0 or 1.}
}
\description{
\code{conditional_p_id} calculates the conditional probabilty of identification
 for a one- or two-stage system given a student's true score, the test reliability,
 and, for two-stage systems, the nomination validity and the nomination cutoff. A
 two-stage system is one in which an initial nomination process is used to select
 students who are tested on the confirmatory assessment.
}
\details{
The probabilities returned by this function can be plotted against a range of true
 scores to create an identification curve (see examples).

The two-stage identification probability is reported if arguments \code{valid}
 and \code{nom.cutoff} are provided. Otherwise, the one-stage probability is
 reported.
}
\examples{
# one-stage system
conditional_p_id(true.score=1, relyt=.9, test.cutoff=.9)

# two-stage system
conditional_p_id(true.score=1, relyt=.9, test.cutoff=.9,
   nom.cutoff=.9, valid=.5)

# make an identification curve
#
# create vector of true scores
Tscores <- seq(0,3, length.out=100)

# calculate the identification probability for each
p.id <- conditional_p_id(true.score=Tscores, relyt=.9,
  test.cutoff=.9, nom.cutoff=.9, valid=.5)

# make a plot
plot(x=Tscores, y=p.id, type="l", xlab="true score",
  ylab="p identified")

# add a reference line for the test cutoff
abline(v=qnorm(.9), col="red")
}
