% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_p_id_2stage_function.R
\name{conditional_p_id_2stage}
\alias{conditional_p_id_2stage}
\title{Function for calculating the conditional probability of identification given
 the true score for a two-stage identification system}
\usage{
conditional_p_id_2stage(relyt, valid, true.score, test.cutoff, nom.cutoff)
}
\arguments{
\item{relyt}{Confirmatory test reliability coefficient. Range (0, 1). 
Must not be exactly 0 or 1.}

\item{valid}{Nomination validity coefficient. Controls the relatedness of the nomination
scores and the confirmatory test scores. Range (0, 1). Must not be exactly 0 or 1, and
must be less than the square root of the test reliability.}

\item{true.score}{The student's true score on a standardized (z-score) metric.}

\item{test.cutoff}{Confirmatory test cutoff percentile. Range (0, 1). 
Must not be exactly 0 or 1.}

\item{nom.cutoff}{Nomination cutoff percentile. Range (0, 1). 
Must not be exactly 0 or 1.}
}
\description{
\code{conditional_p_id_2stage} calculates the conditional probabilty of identification
 for a two-stage system given a student's true score, the test reliability, 
 the nomination validity, and the nomination cutoff. The probabilities returned by 
 this function can be plotted against a range of true scores to create an identification 
 curve.
}
\details{
This function returns the probability of identification.
}
\examples{
# test reliability is .9, the student's true score 
# is 1 SD above the mean, the test cutoff is
# at the 90th percentile, the nomination 
# cutoff is at the 90th percentile, and the
# nomination validity is 0.5
#
conditional_p_id_2stage(relyt=.9, true.score=1, 
   test.cutoff=.9, nom.cutoff=.9, valid=.5)

# make an identification curve
#
# create vector of true scores
T <- seq(0,3, length.out=100)

# calculate the identification probability for each 
p.id <- sapply(T, conditional_p_id_2stage, relyt=.9, 
  test.cutoff=.9, nom.cutoff=.9, valid=.5)

# make a plot
plot(x=T, y=p.id, type="l", xlab="true score", 
  ylab="p identified") 

# add a reference line for the test cutoff
abline(v=qnorm(.9), col="red")
}
