% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_mean_function.R
\name{cor_mean}
\alias{cor_mean}
\title{Correlation between each assessment and the weighted mean of the assessments}
\usage{
cor_mean(r, w = NA)
}
\arguments{
\item{r}{Either a correlation matrix or a vector of unique correlations. If the weights
are not all equal, it is recommended to specify the correlations as a matrix to avoid
erronous pairings of assessment correlations and weights, since this can be confusing
if the correlations are supplied as a vector.}

\item{w}{A vector of weights. Will be internally normalized to sum to 1 and presumes the
same order of assessments as the correlation matrix. If omitted, it is assumed that all
assessments have the same weight.}
}
\description{
This function calculates the correlations between each asssessment and the
 weighted mean of the assessments. If each consistuent assessment is a
 potential basis for nomination, and the weighted mean of all of the assessments
 constitutes the confirmatory test, the returned correlations can be interpreted
 as the nomination validity coefficients that would result from selecting each
 instrument for nomination.
}
\examples{
cor_mean(r=c(.4, .7, .9), w=c(1,2,3))

cor_mean(r=matrix(c( 1,.4,.7,
                      .4, 1,.9,
                      .7,.9, 1), 3,3, byrow=TRUE))

}
