% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_psychometrics_MC_function.R
\name{cross_corrs_lower_left}
\alias{cross_corrs_lower_left}
\title{cross_corrs_lower_left: calculates the cross correlation between true
  scores (rows) and their observed score counterparts (cols). Creates the submatrix
  which occupies the lower left block of the full correlation matrix}
\usage{
cross_corrs_lower_left(corr, rely)
}
\arguments{
\item{corr}{a matrix of observed correlations}

\item{rely}{a vector of reliability coefficients}
}
\value{
a correlation matrix of disattenuated cross-correlations
}
\description{
cross_corrs_lower_left: calculates the cross correlation between true
  scores (rows) and their observed score counterparts (cols). Creates the submatrix
  which occupies the lower left block of the full correlation matrix
}
