% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R,
%   R/marginal_psychometrics_MC_function.R
\name{cross_corrs_upper_right}
\alias{cross_corrs_upper_right}
\title{cross_corrs_upper_right: calculates the cross correlation between observed
  scores (rows) and their true score counterparts (cols). Creates the submatrix
  which occupies the upper right block of the full correlation matrix}
\usage{
cross_corrs_upper_right(corr, rely)

cross_corrs_upper_right(corr, rely)
}
\arguments{
\item{corr}{a matrix of observed correlations}

\item{rely}{a vector of reliability coefficients}
}
\value{
a correlation matrix of disattenuated cross-correlations

a correlation matrix of disattenuated cross-correlations
}
\description{
cross_corrs_upper_right: calculates the cross correlation between observed
  scores (rows) and their true score counterparts (cols). Creates the submatrix
  which occupies the upper right block of the full correlation matrix

cross_corrs_upper_right: calculates the cross correlation between observed
  scores (rows) and their true score counterparts (cols). Creates the submatrix
  which occupies the upper right block of the full correlation matrix
}
