% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errortrapping_function.R
\name{errortrapping}
\alias{errortrapping}
\title{Catch out-of-range parameter values}
\usage{
errortrapping(
  relyt = 1,
  valid = 1e-30,
  nom.cutoff = 0.1,
  test.cutoff = 0.9,
  mu = 0
)
}
\arguments{
\item{relyt}{Confirmatory test reliability coefficient. Range (0, 1].
Must not be exactly 0.}

\item{valid}{Nomination validity coefficient. Range (0, 1).
Must not be exactly 0 or 1.}

\item{nom.cutoff}{Nomination cutoff percentile. Range (0, 1).
Must not be exactly 0 or 1.}

\item{test.cutoff}{Confirmatory test cutoff percentile. Range (0, 1).
Must not be exactly 0 or 1.}

\item{mu}{Population mean true score on a standardized (z-score) metric.}
}
\description{
\code{errortrapping} catches improper input values.
}
\details{
Internal function used to validate parameters supplied to other functions.
}
