% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parms_function.R
\name{estimate_parms}
\alias{estimate_parms}
\title{Estimate psychometric parameters of an identification system given scores
 of the identified students}
\usage{
estimate_parms(scores, w, id.rate, nom.rate, pop.mean = 0, pop.sd = 1,
  adjust = 0.5)
}
\arguments{
\item{scores}{Confirmatory test reliability coefficient. Range (0, 1).
Must not be exactly 0 or 1.}

\item{w}{Optional argument specifying a vector of row numbers that should be included
in the analysis. Defaults to including all cases.}

\item{id.rate}{The proportion of students who have been identified. Range (0, 1). Must
be less than or equal to \code{nom.rate}.}

\item{nom.rate}{The proportion of students who have been nominated. Range (0, 1). Used to
calculate the nomination cutoff.}

\item{pop.mean}{The known general population mean of the scores. Defaults to 0.}

\item{adjust}{Controls the bandwidth of the density estimator. Defaults to 0.5, which
has been found to perform well in simulation.}

\item{pop.mean}{The known general population standard deviation of the scores.
Defaults to 1.}
}
\description{
\code{estimate_params} estimates the psychometric parameters of an identification system
 given the scores of the identified students. The four parameters are the nomination cutoff
 (\code{nom.cutoff}), the confirmatory test cutoff (\code{test.cutoff}), the test reliability
 (\code{relyt}), and the nomination validity (\code{valid}).
}
\details{
The function uses the Levenburg-Marquardt algorithm to minimize the discrepancy
  between the density of the scores and the theoretical unnormalized density of the
  data. See \code{d_identified} for details. The density is estimated via
  shape-constrained kernel density estimation.

Simulation results suggest a mimimum sample size of n=150 to n=200 for reasonable
  performance.
}
\examples{
# generate some data
set.seed(1)
scores <- r_identified(n=500, test.cutoff=.9, relyt=.93, valid=.6,
  nom.cutoff=.85)

# calculate the identification rate implied by the system parameters
id.rate <- marginal_psychometrics(test.cutoff=.9, relyt=.93, valid=.6,
  nom.cutoff=.85)$identification.rate

# calculate the nomination rate implied by the system parameters
nom.rate <- marginal_psychometrics(test.cutoff=.9, relyt=.93, valid=.6,
  nom.cutoff=.85)$nom.rate

# estimate the system parameters
estimate_parms(scores=scores, id.rate=id.rate, nom.rate=nom.rate)

# estimate the parameters using only the first hundred cases
estimate_parms(scores=scores, id.rate=id.rate, nom.rate=nom.rate, w=1:100)


}
