% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_performance_function.R
\name{estimate_performance}
\alias{estimate_performance}
\title{Estimate psychometric parameters of an identification system given scores
 of the identified students using bootstrapping to provide uncertainty estimates.}
\usage{
estimate_performance(scores, id.rate, nom.rate, reps, pop.mean = 0,
  pop.sd = 1, adjust = 0.5, CI = 0.95, ...)
}
\arguments{
\item{scores}{Confirmatory test reliability coefficient. Range (0, 1).
Must not be exactly 0 or 1.}

\item{id.rate}{The proportion of students who have been identified. Range (0, 1). Must
be less than or equal to \code{nom.rate}.}

\item{nom.rate}{The proportion of students who have been nominated. Range (0, 1). Used to
calculate the nomination cutoff.}

\item{reps}{The number of bootstrap samples.}

\item{pop.mean}{The known general population mean of the scores. Defaults to 0.}

\item{CI}{The confidence limit. Must be between 0 and 1. Defaults to .95.}

\item{...}{optional parameters passed to the boot::boot() function. The \code{ncpus} and
\code{parallel} arguments can be specified to increase performance on multiprocessor
hardware.}

\item{pop.mean}{The known general population standard deviation of the scores.
Defaults to 1.}
}
\description{
\code{estimate_performance} estimates the psychometric parameters of an identification system
 given the scores of the identified students. This function calculates statistics
 using the \code{estimate_parms} function with bootstrapping. The resulting \code{relyt},
 \code{test.cutoff}, and \code{valid} estimates are passed to the \code{marginal_psychometrics}
 function in order to calculate the implied performance statistics.
}
\details{
Simulation results suggest a mimimum sample size of n=150 to n=200 for reasonable
  performance. Further, at least 500 bootstrapped samples are suggested.
  This can take a while.
}
\examples{
# generate some data
set.seed(1)
scores <- r_identified(n=500, test.cutoff=.9, relyt=.93, valid=.6,
  nom.cutoff=.85)

# calculate the identification rate implied by the system parameters
id.rate <- marginal_psychometrics(test.cutoff=.9, relyt=.93, valid=.6,
  nom.cutoff=.85)$identification.rate

# calculate the nomination rate implied by the system parameters
nom.rate <- marginal_psychometrics(test.cutoff=.9, relyt=.93, valid=.6,
  nom.cutoff=.85)$nom.rate

# estimate the system parameters with 200 bootstrapped samples
a <- estimate_performance(scores=scores, id.rate=id.rate,
  nom.rate=nom.rate, reps=200)
a

# plot the results for each parameter
plot(a)

}
