% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_valid_function.R
\name{estimate_valid}
\alias{estimate_valid}
\title{Estimate the nomination validity of an identification system given x
 of the identified students}
\usage{
estimate_valid(x, nom.rate, id.rate, pop.mean = 0, pop.sd = 1, adjust = 1)
}
\arguments{
\item{x}{Numeric vector of observed scores.}

\item{nom.rate}{The proportion of students who have been nominated. Range (0, 1). Used to
calculate the nomination cutoff.}

\item{id.rate}{The proportion of students who have been identified. Range (0, 1). Must
be less than or equal to \code{nom.rate}.}

\item{pop.mean}{The known general population mean of the x. Defaults to 0.}

\item{pop.sd}{The known general population standard deviation of the x.
Defaults to 1.}

\item{adjust}{Controls the bandwidth of the density estimator. Defaults to 1.0, which
has been found to perform well in simulation.}
}
\description{
\code{estimate_valid} estimates the nomination validity of an identification system
 given the observed scores of the identified students, the confirmatory test
 cutoff, the nomination cutoff, and the proportion of students that are
 identified. The test cutoff, if not known, is inferred from the minimum score.
 The nomination cutoff is inferred from the proportion of students who are nominated
 under the assumption that the nomination x follow a standard normal distribution.
}
\details{
The function uses the Levenburg-Marquardt algorithm to minimize the discrepancy
  between the density of the x and the theoretical unnormalized density of the
  data. See \code{d_identified} for details.
}
\examples{
# generate some observed scores
# (note the lack of a relyt argument)
# true validity is .6
set.seed(1)
x <- r_identified(n=500, test.cutoff=.9, valid=.6,
  nom.cutoff=.85)

# calculate the identification rate implied by the system parameters
id.rate <- marginal_psychometrics(test.cutoff=.9, valid=.6,
  nom.cutoff=.85)$identification.rate

# calculate the nomination rate implied by the system parameters
nom.rate <- marginal_psychometrics(test.cutoff=.9, valid=.6,
  nom.cutoff=.85)$nom.rate

# estimate the system parameters from the data
estimate_valid(x=x, id.rate=id.rate, nom.rate=nom.rate)

}
