% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{giftedCalcs-package}
\alias{giftedCalcs}
\alias{giftedCalcs-package}
\title{Psychometrics for gifted and talented program identification}
\description{
This package contains a suite of functions for exploring the psychometrics of one- and
  two-stage identification systems. A set of functions for understanding the conditional
  distribution of ability scores for identified students are available. Further, given
  a set of scores from identified students, program characteristics such as sensitivity
  and the incorrect identification rate (e.g., 1-PPV) can be estimated with bootstrapped
  uncertainty.
}
\details{
For a complete list of functions, use \code{library(help="giftedCalcs")}
}
\author{
\strong{Maintainer}: Matthew McBee \email{mcbeem@etsu.edu}

}
