% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_psychometrics_function.R
\name{marginal_psychometrics}
\alias{marginal_psychometrics}
\title{Marginal performance statistics identification systems}
\usage{
marginal_psychometrics(relyt, test.cutoff, valid = 1e-07,
  nom.cutoff = 1e-07, mu = 0)
}
\arguments{
\item{relyt}{Confirmatory test reliability coefficient. Range (0, 1).
Must not be exactly 0 or 1.}

\item{test.cutoff}{Confirmatory test cutoff percentile. Range (0, 1).
Must not be exactly 0 or 1.}

\item{valid}{Nomination validity coefficient. Controls the relatedness of the nomination
scores and the confirmatory test scores. Range (0, 1). Must not be exactly 0 or 1, and
must be less than the square root of the test reliability.}

\item{nom.cutoff}{Nomination cutoff percentile. Range (0, 1).
Must not be exactly 0 or 1.}

\item{mu}{Population mean true score on a standardized (z-score) metric.
Defaults to zero.}
}
\description{
This function calculates marginal performance statistics for
 one- and two-stage identification systems. A two-stage system is one in which
 an initial nomination process is used to select students who are tested on
 the confirmatory assessment.
}
\details{
Two-stage system results are reported if arguments valid and nom.cutoff
 are provided. Otherwise, one-stage results are reported.

The function returns a list containing the following:

sensitivity: The proprtion of qualifying students who are identified.

IIR: Incorrect identification rate; the proportion of identified students
   who do not qualify.

nom.rate: Nomination rate; the proportion of students who are nominated, which controls
   testing costs. (Two-stage only).

nom.passrate: The proportion of nominated students who go on to be identified.
   (Two-stage only)

identification.rate: The proportion of the student population that is identified.
}
\examples{
# one-stage system
marginal_psychometrics(mu=0, test.cutoff=.9, relyt=.95)

# two-stage system
marginal_psychometrics(mu=0, test.cutoff=.9,
        nom.cutoff=.8, relyt=.95, valid=.7)

}
