% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_psychometrics_2stage_function.R
\name{marginal_psychometrics_2stage}
\alias{marginal_psychometrics_2stage}
\title{Marginal performance statistics for a two-stage
 identification system.}
\usage{
marginal_psychometrics_2stage(relyt, valid, nom.cutoff, test.cutoff, mu = 0)
}
\arguments{
\item{relyt}{Confirmatory test reliability coefficient. Range (0, 1).
Must not be exactly 0 or 1.}

\item{valid}{Nomination validity coefficient. Controls the relatedness of the nomination
scores and the confirmatory test scores. Range (0, 1). Must not be exactly 0 or 1, and
must be less than the square root of the test reliability.}

\item{nom.cutoff}{Nomination cutoff percentile. Range (0, 1).
Must not be exactly 0 or 1.}

\item{test.cutoff}{Confirmatory test cutoff percentile. Range (0, 1).
Must not be exactly 0 or 1.}

\item{mu}{Population mean true score on a standardized (z-score) metric.
Defaults to zero.}
}
\description{
\code{marginal_psychometrics_2stage} calculates marginal performance statistics
 for a two-stage identification system, in which an initial nomination process
 is used to select students who are tested on the confirmatory assessment.
}
\details{
This function returns a list containing the following:

sensitivity: The proprtion of qualifying students who are identified.

IIR: Incorrect identification rate; the proportion of identified students
   who do not qualify.

nom.rate: Nomination rate; the proportion of students who are nominated, which controls
   testing costs.

nom.passrate: The proportion of nominated students who go on to be identified.

identification.rate: The proportion of the student population that is identified.
}
\examples{
# the population mean ability is zero, the test cutoff is
# at the 90th percentile,the nomination cutoff is at the 80th
# percentile, the confirmatory test reliability
# is .95, and the nomination validity is .7

 marginal_psychometrics_2stage(mu=0, test.cutoff=.9,
  nom.cutoff=.8, relyt=.95, valid=.7)

}
