% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_psychometrics_MC_function.R
\name{marginal_psychometrics_MC}
\alias{marginal_psychometrics_MC}
\title{Marginal psychometrics for multiple criteria identification systems}
\usage{
marginal_psychometrics_MC(policy, corr, rely, n=50000, nomination=NA,
  ignore_nomination=FALSE, labels=NA)
}
\arguments{
\item{policy}{a matrix describing the identification policy. assessments are in
columns, pathways in are in rows. values are percentile cutoffs. within a row,
multiple requirements are joined by "and" combination rules, whereas the "or"
rule joins across rows}

\item{corr}{a correlation matrix}

\item{rely}{a vector of reliability coefficients}

\item{n}{scalar, the number of samples to draw. defaults to 50,000}

\item{nomination}{scalar defining which column of the policy matrix,
row / column of the correlation matrix, and element of the reliability
vector is the nomination. Defaults to NA, which is interpreted as no
nomination stage}

\item{ignore_nomination}{boolean. Should the nomination be ignored? This allows
for convenient comparison of single- and two-stage versions of a multiple
criteria policy without needing to respecify the other inputs. defaults to FALSE}

\item{labels}{an optional vector of labels for the assessments; defaults to NA}
}
\value{
an object of class giftedCalcsMC, a list with the following elements:
   $identified: the proportion of students that are identified
   $gifted: the proportion of students that are gifted
   $sensitivity: the sensitivity
   $scores: a data of scores for identified students
}
\description{
This function calculates marginal psychometrics (currently only sensitivity) for
one- and two-stage multiple criteria identification systems. A plot method is
available for visualizing the distributions of scores on each assessment for the
identified student. This function uses Monte Carlo simulation to approximate the
integrals involved in computing the metrics; as such the results will vary somewhat
from run to run. Users should determine the necessary sample size (\code{n}) needed
to achieve the required level of precision. The default value of \code{n=50000} will
likely suffice for most situations. Users may also wish to set a random number seed
for reproducibility.
}
\examples{

policy = matrix(c(
 .9, .9, .9, 0,
 .9, 0, .9, .9,
 .9, 0, 0, .95), ncol=4, byrow=T)

corr = matrix(c(  1,  .5,  .4, .3,
                 .5,  1, .7, .6,
                 .4,  .7,  1, .5,
                 .3,  .6, .5,  1), byrow=T, nrow=4)

rely=c(.8, .9, .8, .85)

result = marginal_psychometrics_MC(n=50000, policy=policy, corr=corr,
 rely=rely, nomination=1, labels=c("nom", "IQ", "ach", "creativity"),
 ignore_nomination=F)

result

plot(result)

}
