% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_identified_function.R
\name{mean_identified}
\alias{mean_identified}
\title{Mean of the conditional distribution of true or observed scores for identified students}
\usage{
mean_identified(
  relyt = 1,
  test.cutoff,
  valid = 1e-07,
  nom.cutoff = 1e-07,
  mu = 0
)
}
\arguments{
\item{relyt}{Confirmatory test reliability coefficient. Range (0, 1].
Must not be exactly 0. Defaults to 1; in this case, x is assumed
to be an observed score. If an alternative value is supplied for
\code{relyt}, x is assumed to be a true score.}

\item{test.cutoff}{Confirmatory test cutoff percentile. Range (0, 1).
Must not be exactly 0 or 1.}

\item{valid}{Nomination validity coefficient. Controls the relatedness of the nomination
scores and the confirmatory test scores. Range (0, 1). Must not be exactly 0 or 1, and
must be less than the square root of the test reliability. Defaults to 1e-7 for a single-
stage identification system.}

\item{nom.cutoff}{Nomination cutoff percentile. Range (0, 1).
Must not be exactly 0 or 1. Defaults to 1e-7 for a single-
stage identification system.}

\item{mu}{Population mean true score on a standardized (z-score) metric.
Defaults to zero.}
}
\description{
This function calculates the expected value (mean) of the ability true
  score distribution for identified students.
}
\examples{
# mean true score for identified students
mean_identified(relyt=.9, valid=.8,
  test.cutoff=.9, nom.cutoff=.5)

# mean observed score for identified students
mean_identified(valid=.8,test.cutoff=.9,
nom.cutoff=.5)
}
