% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_est_performance_method.R
\name{plot.est_performance}
\alias{plot.est_performance}
\title{Plot method for class est_performance}
\usage{
\method{plot}{est_performance}(x, width = 0.3, adjust = 1.2,
  type = "density", ...)
}
\arguments{
\item{width}{Adjusts the scale of the x-axis. Defaults to 0.3.}

\item{adjust}{Adjusts the smoothing bandwidth for density estimation. Defaults to 1.2.}

\item{type}{A character string selecting density plots or histograms. Must be one of
"density" or "hist"}
}
\description{
Plot method for class est_performance
}
