% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_est_performance_method.R
\name{print.est_performance}
\alias{print.est_performance}
\title{Print method for class est_performance}
\usage{
\method{print}{est_performance}(x, ...)
}
\arguments{
\item{x}{an object of class est_performance to be printed}

\item{...}{additional arguments to be passed to \code{print}}
}
\description{
Print method for class est_performance
}
