% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_psychometrics_MC_function.R
\name{print.giftedCalcsMC}
\alias{print.giftedCalcsMC}
\title{print.giftedCalcsMC: print method for class giftedCalcsMC
suppress the printing of scores for identified students}
\usage{
\method{print}{giftedCalcsMC}(x, ...)
}
\arguments{
\item{x}{a list, the output of marginal_psychometrics_MC}

\item{..., }{additional arguments passed to print()}
}
\value{
a printed subset of x
}
\description{
print.giftedCalcsMC: print method for class giftedCalcsMC
suppress the printing of scores for identified students
}
