% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_perf_by_cutoff_method.R
\name{print.perf_by_cutoff}
\alias{print.perf_by_cutoff}
\title{Print method for class perf_by_cutoff}
\usage{
\method{print}{perf_by_cutoff}(x, ...)
}
\arguments{
\item{x}{an object of class perf_by_cutoff to be printed}

\item{...}{additional arguments to be passed to \code{print}}
}
\description{
Print method for class perf_by_cutoff
}
