% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability_mean_function.R
\name{reliability_mean}
\alias{reliability_mean}
\title{Reliability of the weighted mean of assessments}
\usage{
reliability_mean(rely, r, w = NA)
}
\arguments{
\item{rely}{A vector of reliability coefficients.}

\item{r}{Either a correlation matrix or a vector of unique correlations. It is
recommended to specify the correlations as a matrix to avoid
erronous pairings of assessment correlations with reliability coefficients and weights,
since this can be confusing if the correlations are supplied as a vector. Presumes that the
correlation matrix has the same order of assessments as the reliability and weight vectors.}

\item{w}{A vector of weights. Will be internally normalized to sum to 1 and presumes the
same order of assessments as the correlation matrix and vector of reliabilities.
If omitted, it is assumed that all assessments have the same weight.}
}
\description{
This function calculates the reliability of a weighted mean of assessments given the
 reliability coefficient of each assessment, the correlations between them, and the
 weights.
}
\examples{
r <- matrix(c(
  1, .4, .7,
  .4, 1, .5,
  .7, .5, 1
), 3, 3, byrow = TRUE)
reliability_mean(rely = c(.9, .85, .88), r = r, w = c(.5, .3, .2))
}
