% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_identified_function.R
\name{sd_identified}
\alias{sd_identified}
\title{Function for calculing the standard deviation of the conditional distribution of ability for identified students}
\usage{
sd_identified(relyt, valid, test.cutoff, nom.cutoff, mu = 0)
}
\arguments{
\item{relyt}{Confirmatory test reliability coefficient. Range (0, 1).
Must not be exactly 0 or 1.}

\item{valid}{Nomination validity coefficient. Controls the relatedness of the nomination
scores and the confirmatory test scores. Range (0, 1). Must not be exactly 0 or 1, and
must be less than the square root of the test reliability.}

\item{test.cutoff}{Confirmatory test cutoff percentile. Range (0, 1).
Must not be exactly 0 or 1.}

\item{nom.cutoff}{Nomination cutoff percentile. Range (0, 1).
Must not be exactly 0 or 1.}

\item{mu}{Population mean true score on a standardized (z-score) metric.
Defaults to zero.}
}
\description{
\code{sd_identified} calculates the expected value (mean) of the ability true
  score distribution for identified students.
}
\examples{
# std dev on the z-score metric
sd_identified(relyt=.9, valid=.8,
  test.cutoff=.9, nom.cutoff=.5, mu=0)

# sd on the IQ-metric
15*(sd_identified(relyt=.9, valid=.8,
test.cutoff=.9, nom.cutoff=.5, mu=0))
}
