% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R,
%   R/marginal_psychometrics_MC_function.R
\name{true_corrs}
\alias{true_corrs}
\title{true_corrs: calculates the true score correlation matrix from an observed score
  correlation matrix}
\usage{
true_corrs(corr, rely)

true_corrs(corr, rely)
}
\arguments{
\item{corr}{an n by n observed correlation matrix}

\item{rely}{a vector of reliability coefficients}
}
\value{
a correlation matrix of true scores which are disattentuated for
  measurement error

a correlation matrix of true scores which are disattentuated for
  measurement error
}
\description{
true_corrs: calculates the true score correlation matrix from an observed score
  correlation matrix

true_corrs: calculates the true score correlation matrix from an observed score
  correlation matrix
}
