% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_moments_function.R
\name{conditional_moments}
\alias{conditional_moments}
\title{Conditional moments of scores}
\usage{
conditional_moments(t.true = NA, n.obs = NA, t.obs = NA, relyt, valid = 1e-07)
}
\arguments{
\item{t.true}{The confirmatory test true score on a standardized (z-score) metric. Only
one of \code{t.true}, \code{t.obs}, or \code{n.obs} should be specified}

\item{n.obs}{The nomination observed score on a standardized (z-score) metric. If \code{n.obs}
is specified, a value for \code{valid} must also be given.}

\item{t.obs}{The confirmatory test observed score on a standardized (z-score) metric.}

\item{relyt}{Confirmatory test reliability coefficient. Range (0, 1).
Must not be exactly 0 or 1.}

\item{valid}{Nomination validity coefficient. Controls the relatedness of the nomination
 scores and the confirmatory test scores. Range (0, 1). Must not be exactly 0 or 1, and
 must be less than the square root of the test reliability. If provided, the two-stage
 version of the computation is performed.

 @return A list containing the following:

 conditional.mean: For one-stage systems, the mean of the confirmatory test scores.
 For two-stage systems, the mean vector of the bivariate normal distribution of
 the observed nomination and confirmatory test scores, in that order.

 conditional.cov: For one-stage systems, the variance of the observed confirmatory
  test scores. For two-stage systems, the covariance matrix of the bivariate
  normal distribution of the observed nomination and confirmatory test scores,
  in that order.}
}
\description{
This function calculates the conditional mean and variance of scores. The user specifies
 a value for the confirmatory test true score, the confirmatory test observed score, or
 the nomination observed score. The moments of the conditional distribution of the other
 two scores are calculated.
}
\details{
In a two-stage identification system, there is a test true score, a test observed score,
 and a nomination observed score. (In principle there is also a nomination true score, but
 it is irrelevant and may be ignored). In a one-stage system, there is a test true score and
 observed score.

The joint distribution of these three variables is assumed to be multivariate
normal. Therefore, the conditional distribution is also normal.

If the argument valid is supplied, the two-stage conditional moments are
 returned. Otherwise the one-stage moments are returned.
}
\examples{

# one-stage system
conditional_moments(t.true = 2, relyt = .9)

# two-stage system
conditional_moments(t.true = 2, relyt = .9, valid = .6)
}
