% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_identified.R
\name{d_identified}
\alias{d_identified}
\title{Function for the conditional density of true scores for identified students}
\usage{
d_identified(relyt, valid, test.cutoff, nom.cutoff, true.score, mu = 0)
}
\arguments{
\item{relyt}{Confirmatory test reliability coefficient. Range (0, 1). 
Must not be exactly 0 or 1.}

\item{valid}{Nomination validity coefficient. Controls the relatedness of the nomination
scores and the confirmatory test scores. Range (0, 1). Must not be exactly 0 or 1, and
must be less than the square root of the test reliability.}

\item{test.cutoff}{Confirmatory test cutoff percentile. Range (0, 1). 
Must not be exactly 0 or 1.}

\item{nom.cutoff}{Nomination cutoff percentile. Range (0, 1). 
Must not be exactly 0 or 1.}

\item{true.score}{The student's true score on a standardized (z-score) metric.}

\item{mu}{Population mean true score on a standardized (z-score) metric.
Defaults to zero.}
}
\description{
\code{d_identified} is the conditional probability density function (pdf) for 
identified students. It is a true probablity density with a total area of one.
}
\details{
See also \code{d_identified_unnormed} for the unnormalized density, \code{p_identified} 
for the cumulative density, \code{q_identified} for the quantile function, and 
\code{r_identified} for random generation.
}
\examples{
# test reliability is .9, the student's true score 
# is 1 SD above the mean, the test cutoff is
# at the 90th percentile, the nomination 
# cutoff is at the 90th percentile, the
# nomination validity is 0.5, and the student's
# population mean ability true score is 0
#
d_identified(relyt=.9, true.score=1, test.cutoff=.9, 
  nom.cutoff=.9, valid=.5, mu=0)

# make plot of the density
#
# create vector of true scores
T <- seq(0,4, length.out=200)

# plot the density
p.id <- sapply(T, d_identified, relyt=.95, 
  test.cutoff=.9, nom.cutoff=.9, valid=.5)

plot(x=T, y=p.id, type="l", xlab="true score", ylab="density") 
}
