% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_identified_function.R
\name{mean_identified}
\alias{mean_identified}
\title{Mean of the conditional distribution of ability for identified students}
\usage{
\code{mean_identified(relyt, test.cutoff, valid, nom.cutoff, mu=0)}
}
\arguments{
\item{relyt}{Confirmatory test reliability coefficient. Range (0, 1).
Must not be exactly 0 or 1.}

\item{test.cutoff}{Confirmatory test cutoff percentile. Range (0, 1).
Must not be exactly 0 or 1.}

\item{valid}{Nomination validity coefficient. Controls the relatedness of the nomination
scores and the confirmatory test scores. Range (0, 1). Must not be exactly 0 or 1, and
must be less than the square root of the test reliability.}

\item{nom.cutoff}{Nomination cutoff percentile. Range (0, 1).
Must not be exactly 0 or 1.}

\item{mu}{Population mean true score on a standardized (z-score) metric.
Defaults to zero.}
}
\description{
\code{mean_identified} calculates the expected value (mean) of the ability true
  score distribution for identified students.
}
\examples{
# mean on the z-score metric
mean_identified(relyt=.9, valid=.8,
  test.cutoff=.9, nom.cutoff=.5, mu=0)

# mean on the IQ-metric
100+(mean_identified(relyt=.9, valid=.8,
test.cutoff=.9, nom.cutoff=.5, mu=0)*15)
}
