% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbboot_function.R
\name{pbboot}
\alias{pbboot}
\title{Bootstrapping with a progress bar}
\usage{
pbboot(data, statistic, R, sim = "ordinary", stype = c("i", "f", "w"),
  strata = rep(1, n), L = NULL, m = 0, weights = NULL,
  ran.gen = function(d, p) d, mle = NULL, simple = FALSE, ...,
  parallel = c("no", "multicore", "snow"), ncpus = getOption("boot.ncpus",
  1L), cl = NULL)
}
\description{
This function reproduces the code from the \code{boot} package's
 \code{boot()} function, changing one line in order to add a progress
 bar using \code{pbapply} package. This code was adopted from a response
 posted to StackOverflow by user psolymos.
 See http://goo.gl/F526BQ for details and \code{?boot::boot} for help.
}
\details{
@param data Data to for resampling
 @param statistic Function returning estimates when applied to data
 @param R number of bootstrap repetitions
 @parallel optional; a character string specifying the package to use for parallel
   processing, must be one of "snow", "multicore", or "no"
 @ncpus optional; number of CPUs

@examples

m.boot <- function(data, indices) {
  d <- data[indices]
  mean(d, na.rm = T)
  }

tot_rep <- 200
 v1 <- rnorm(1000)
 b <- pbboot(v1, m.boot, R = tot_rep)
}
