% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R,
%   R/marginal_psychometrics_MC_function.R
\name{plot.giftedCalcsMC}
\alias{plot.giftedCalcsMC}
\title{plot.giftedCalcsMC: plot method}
\usage{
\method{plot}{giftedCalcsMC}(x)

\method{plot}{giftedCalcsMC}(x)
}
\arguments{
\item{x}{list, the output of the marginal_psychometrics_MC function}
}
\value{
a ggplot ridgeline plot

a ggplot ridgeline plot
}
\description{
plot.giftedCalcsMC: plot method

plot.giftedCalcsMC: plot method
}
