% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_est_performance_method.R
\name{print.est_performance}
\alias{print.est_performance}
\title{Print method for class est_performance}
\usage{
\method{print}{est_performance}(x, ...)
}
\description{
Print method for class est_performance
}
