% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_identified.R
\name{q_identified}
\alias{q_identified}
\title{Conditional quantile function for the distribution of true scores for identified students}
\usage{
q_identified(relyt, valid, test.cutoff, nom.cutoff, percentile, mu = 0)
}
\arguments{
\item{relyt}{Confirmatory test reliability coefficient. Range (0, 1). 
Must not be exactly 0 or 1.}

\item{valid}{Nomination validity coefficient. Controls the relatedness of the nomination
scores and the confirmatory test scores. Range (0, 1). Must not be exactly 0 or 1, and
must be less than the square root of the test reliability.}

\item{test.cutoff}{Confirmatory test cutoff percentile. Range (0, 1). 
Must not be exactly 0 or 1.}

\item{nom.cutoff}{Nomination cutoff percentile. Range (0, 1). 
Must not be exactly 0 or 1.}

\item{percentile}{The percentile of the distribution of identified students. Range (0, 1).
Must not be exactly 0 or 1.}

\item{mu}{Population mean true score on a standardized (z-score) metric.
Defaults to zero.}
}
\description{
\code{q_identified} is the conditional quantile function for identified students. 
Given a percentile for the true score, it returns the corresponding true score (on a
z-score metric).
}
\details{
See also \code{d_identified_unnormed} for the unnormalized density, \code{d_identified} 
for the normalized density, \code{p_identified} for the cumulative density function, and 
\code{r_identified} for random generation.
}
\examples{
q_identified(percentile=.9, relyt=.95, valid=.6, 
  test.cutoff=.975, nom.cutoff=.9, mu=0)
}
