% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_identified_function.R
\name{sd_identified}
\alias{sd_identified}
\title{Standard deviation of the conditional distribution of true or observed scores for
 identified students}
\usage{
sd_identified(
  relyt = 1,
  test.cutoff,
  valid = 1e-07,
  nom.cutoff = 1e-07,
  mu = 0
)
}
\arguments{
\item{relyt}{Confirmatory test reliability coefficient. Range (0, 1].
Must not be exactly 0. Defaults to 1; in this case, x is assumed
to be an observed score. If an alternative value is supplied for
\code{relyt}, x is assumed to be a true score.}

\item{test.cutoff}{Confirmatory test cutoff percentile. Range (0, 1).
Must not be exactly 0 or 1.}

\item{valid}{Nomination validity coefficient. Controls the relatedness of the nomination
scores and the confirmatory test scores. Range (0, 1). Must not be exactly 0 or 1, and
must be less than the square root of the test reliability.}

\item{nom.cutoff}{Nomination cutoff percentile. Range (0, 1).
Must not be exactly 0 or 1.}

\item{mu}{Population mean true score on a standardized (z-score) metric.
Defaults to zero.}
}
\description{
\code{sd_identified} calculates the standard deviation of the ability true
  score distribution for identified students.
}
\examples{
# std dev for true scores
sd_identified(relyt=.9, valid=.8,
  test.cutoff=.9, nom.cutoff=.5, mu=0)

# std dev for observed scores
15*(sd_identified(valid=.8,
test.cutoff=.9, nom.cutoff=.5, mu=0))
}
